style: unwrap single-item Capability.contributes returns; move debug types out of barrel index
- Remove unnecessary array wrapper from single Capability.contributes returns in plugin-assistant, plugin-automation, plugin-inbox, plugin-thread app-graph-builder modules
- Move DebugPluginOptions, DebugContextType, DebugContext, DebugCapabilities from types/index.ts to types/debug.ts so index.ts is barrel-only
