+++ packages/plugins/plugin-assistant/src/capabilities/app-graph-builder.ts
@@ -239,6 +239,6 @@ export default Capability.makeModule(
       }),
     ]);
 
-    return [Capability.contributes(AppCapabilities.AppGraphBuilder, extensions)];
+    return Capability.contributes(AppCapabilities.AppGraphBuilder, extensions);
   }),
 );
