+++ packages/plugins/plugin-automation/src/capabilities/app-graph-builder.ts
@@ -52,6 +52,6 @@ export default Capability.makeModule(
       }),
     ]);
 
-    return [Capability.contributes(AppCapabilities.AppGraphBuilder, extensions)];
+    return Capability.contributes(AppCapabilities.AppGraphBuilder, extensions);
   }),
 );
