+++ packages/plugins/plugin-debug/src/types/index.ts
@@ -2,38 +2,7 @@
 // Copyright 2023 DXOS.org
 //
 
-import { type Atom } from '@effect-atom/atom-react';
-import { type Context, createContext } from 'react';
-
-import { Capability } from '@dxos/app-framework';
-import type { TimerCallback, TimerOptions } from '@dxos/async';
-import { type IdbLogStore } from '@dxos/log-store-idb';
-
-import { meta } from '#meta';
-import * as Settings from './Settings';
-
+export * from './debug';
 export * from './surface';
 export * as Devtools from './Devtools';
 export * as Settings from './Settings';
-
-export type DebugPluginOptions = {
-  /** Shared persistent log store for capturing and downloading logs. */
-  logStore?: IdbLogStore;
-};
-
-export type DebugContextType = {
-  running: boolean;
-  start: (cb: TimerCallback, options: TimerOptions) => void;
-  stop: () => void;
-};
-
-export const DebugContext: Context<DebugContextType> = createContext<DebugContextType>({
-  running: false,
-  start: () => {},
-  stop: () => {},
-});
-
-export namespace DebugCapabilities {
-  export const Settings = Capability.make<Atom.Writable<Settings.Settings>>(`${meta.profile.key}.capability.settings`);
-}
-
