+++ packages/plugins/plugin-inbox/src/capabilities/app-graph-builder.ts
@@ -607,6 +607,6 @@ export default Capability.makeModule(
       }),
     ]);
 
-    return [Capability.contributes(AppCapabilities.AppGraphBuilder, extensions)];
+    return Capability.contributes(AppCapabilities.AppGraphBuilder, extensions);
   }),
 );
