+++ packages/plugins/plugin-thread/src/capabilities/app-graph-builder.ts
@@ -83,6 +83,6 @@ export default Capability.makeModule(
       }),
     ]);
 
-    return [Capability.contributes(AppCapabilities.AppGraphBuilder, extensions)];
+    return Capability.contributes(AppCapabilities.AppGraphBuilder, extensions);
   }),
 );
